Replaced all shorts with ints (And SMALLINT with INT) to ensure enough member capacity (design oversight...). Added documentation build to build system. Some other slight doc update on the Database interface.

+++ trunk/scripts/simpleDB/setupMembers.sql
@@ -24,10 +24,10 @@
 # Script: setupMembers
 # Requires: groups
 CREATE TABLE members (
-    id SMALLINT NOT NULL GENERATED ALWAYS AS IDENTITY CONSTRAINT MEMBERS_PK PRIMARY KEY,
+    id INT NOT NULL GENERATED ALWAYS AS IDENTITY CONSTRAINT MEMBERS_PK PRIMARY KEY,
     surname VARCHAR(32),
     forename VARCHAR(32),
     dob DATE NOT NULL,
-    usergroup SMALLINT,
+    usergroup INT,
     CONSTRAINT MEMBERS_GROUPS_FK FOREIGN KEY (usergroup) references groups (id)
 )
